combined heap_scan and scan
